Add DCOM, SupportsShouldProcess support
Updated it to support multiple computers and multiple users, added an option to turn on DCOM for the CIM session if requested.  
Changed the CIM call to filter the user profiles in the Get-CimInstance call instead of after pulling them all.  
Cleaned up the help so that the Parameters show correctly.